requirements.txt
asgiref==3.8.1
Django==5.1.1
django-cors-headers==4.4.0
django-formtools==2.5.1
django-otp==1.5.4
django-phonenumber-field==8.0.0
django-two-factor-auth==1.17.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
phonenumbers==8.13.46
pillow==10.4.0
PyJWT==2.9.0
pypng==0.20220715.0
PyYAML==6.0.2
qrcode==7.4.2
sqlparse==0.5.1
typing_extensions==4.12.2
ua-parser==0.18.0
user-agents==2.2.0
